Refactored basis.type tests
